add .pmtignore based on Vox Pupuli
this file will be read if you run "puppet module generate". This is
needed if you ever want to upload modules to forge.puppet.com. If this
file is nonexistent the .gitignore will be read. There are files you may
want to track but not to release, for example documentation